% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chacko63_tab1}
\alias{chacko63_tab1}
\title{Table I}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10 rows and 8 columns.
}
\usage{
chacko63_tab1
}
\description{
Table of \eqn{p_{m, k}}. Gives the values of \eqn{p_{m, k}} for
equal sample sizes and \eqn{k = 3, 4, \ldots, 10}.
}
\references{
Chacko, V. J. (1963). Testing homogeneity against ordered alternatives. The
Annals of Mathematical Statistics, 945-956.
}
\keyword{datasets}
