% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chacko66_sec3}
\alias{chacko66_sec3}
\title{Chacko (1966), section 3}
\format{
An object of class \code{integer} of length 5.
}
\usage{
chacko66_sec3
}
\description{
A multinomial example with 5 cell frequencies
}
\references{
Chacko, V. J. (1966). Modified chi-square test for ordered alternatives.
Sankhyā: The Indian Journal of Statistics, Series B, 185-190.
}
\keyword{datasets}
