% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permChacko-package.R
\docType{package}
\name{permChacko-package}
\alias{permChacko}
\alias{permChacko-package}
\title{permChacko: Chacko Test for Order-Restriction with Permutation}
\description{
Implements an extension of the Chacko chi-square test for ordered vectors (Chacko, 1966, \url{https://www.jstor.org/stable/25051572}). Our extension brings the Chacko test to the computer age by implementing a permutation test to offer a numeric estimate of the p-value, which is particularly useful when the analytic solution is not available.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ocbe-uio.github.io/permChacko/}
  \item Report bugs at \url{https://github.com/ocbe-uio/permChacko/issues}
}

}
\author{
\strong{Maintainer}: Waldir Leoncio \email{w.l.netto@medisin.uio.no} (\href{https://orcid.org/0000-0002-6719-6162}{ORCID})

Authors:
\itemize{
  \item Graeme Ruxton
  \item Morten Wang Fagerland \email{morten.fagerland@medisin.uio.no}
}

}
\keyword{internal}
