% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permChacko.R
\name{permChacko}
\alias{permChacko}
\title{The Chacko test for order-restriction with permutation test}
\usage{
permChacko(x)
}
\arguments{
\item{x}{vector of numeric values}
}
\description{
The Chacko test for order-restriction with permutation test
}
