% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ruxton221207}
\alias{ruxton221207}
\title{Example by Graeme Ruxton}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
ruxton221207
}
\description{
A simple example showing how a vector can be reduced to one
element.
}
\keyword{datasets}
