% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permChacko.R
\name{permChacko}
\alias{permChacko}
\title{The Chacko test for order-restriction with permutation test}
\usage{
permChacko(x, verbose = FALSE)
}
\arguments{
\item{x}{vector of numeric values}

\item{verbose}{if \code{TRUE}, prints intermediate messages and output}
}
\description{
The Chacko test for order-restriction with permutation test
}
\references{
Chacko, V. J. (1963). Testing homogeneity against ordered alternatives. The
Annals of Mathematical Statistics, 945-956.

Chacko, V. J. (1966). Modified chi-square test for ordered alternatives.
Sankhyā: The Indian Journal of Statistics, Series B, 185-190.
}
