% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceVector.R
\name{reduceVector}
\alias{reduceVector}
\title{Reduce a vector using the ordering process}
\usage{
reduceVector(x, verbosity = 0L)
}
\arguments{
\item{x}{a vector of numeric values}

\item{verbosity}{a natural number indicating the amount of output to print}
}
\value{
A matrix with two columns: the first one is the reduced vector and
the second one is its weight
}
\description{
This function implements the ordering process described in
Chacko (1963) and Chacko (1966).
}
\examples{
reduceVector(c(10, 16, 14, 12, 18))
reduceVector(c(10, 8, 4, 2, 1))
reduceVector(chacko66_3)
reduceVector(chacko66_5)
}
\author{
Waldir Leoncio
}
